refactor: move presets into base-testing-tool
