+++ package.json
@@ -39,7 +39,6 @@
   "workspaces": [
     "libraries/*",
     "plugins/*",
-    "presets/*",
     "tools/*"
   ],
   "devDependencies": {
