--- presets/preset-dynamosa/package.json
@@ -1,55 +1,0 @@
-{
-  "name": "@syntest/preset-dynamosa",
-  "version": "0.0.0",
-  "description": "The DynaMOSA preset of the SynTest Framework",
-  "keywords": [
-    "syntest",
-    "automatic-test-generation"
-  ],
-  "homepage": "https://www.syntest.org",
-  "bugs": {
-    "url": "https://github.com/syntest-framework/syntest-core/issues"
-  },
-  "license": "Apache-2.0",
-  "contributors": [
-    "Annibale Panichella",
-    "Mitchell Olsthoorn",
-    "Dimitri Stallenberg"
-  ],
-  "main": "dist/index.js",
-  "types": "dist/index.d.ts",
-  "files": [
-    "dist",
-    "NOTICE",
-    "README.md"
-  ],
-  "repository": {
-    "type": "git",
-    "url": "git+https://github.com/syntest-framework/syntest-core.git"
-  },
-  "scripts": {
-    "build": "tsc --build",
-    "build:watch": "tsc --build --watch",
-    "clean": "rm -rf .nyc_output dist node_modules coverage",
-    "clean:dist": "rm -rf dist",
-    "format": "prettier --config ../../.prettierrc.json --ignore-path ../../.prettierignore --write .",
-    "format:check": "prettier --config ../../.prettierrc.json --ignore-path ../../.prettierignore --check .",
-    "lint": "eslint --config ../../.eslintrc.json --ignore-path ../../.eslintignore .",
-    "lint:fix": "eslint --config ../../.eslintrc.json --ignore-path ../../.eslintignore . --fix",
-    "test": "mocha --config ../../.mocharc.json",
-    "test:coverage": "nyc --reporter=text --reporter=html mocha --config ../../.mocharc.json",
-    "test:coverage:ci": "nyc --reporter=lcovonly mocha --config ../../.mocharc.json --reporter json --reporter-option output=test-results.json",
-    "test:watch": "mocha --config ../../.mocharc.json --watch"
-  },
-  "dependencies": {
-    "@syntest/base-testing-tool": "^0.2.0-beta.9",
-    "@syntest/logging": "*",
-    "@syntest/module": "*"
-  },
-  "engines": {
-    "node": ">=10.24.0"
-  },
-  "publishConfig": {
-    "access": "public"
-  }
-}
