--- presets/preset-mosa/NOTICE
@@ -1,5 +1,0 @@
-SynTest Framework - SynTest Core
-Copyright 2020-2023 Delft University of Technology and SynTest contributors
-
-This product includes software developed at
-Delft University of Technology (http://www.tudelft.nl/).
