--- presets/preset-mosa/lib/tsconfig.json
@@ -1,9 +1,0 @@
-{
-  "extends": "../../../tsconfig-base.json",
-  "compilerOptions": {
-    "outDir": "../dist",
-    "rootDir": ".",
-    "composite": true
-  },
-  "include": ["**/*.ts"]
-}
