--- presets/preset-mosa/tsconfig.json
@@ -1,6 +1,0 @@
-{
-  "extends": "../../tsconfig-base.json",
-  "files": [],
-  "include": [],
-  "references": [{ "path": "lib/" }]
-}
