--- presets/preset-nsgaii/README.md
@@ -1,1 +1,0 @@
-# The init application of the syntest ecosystem
